# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/users":{"get":{"operationId":"UserController_init","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/register":{"post":{"operationId":"UserController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/users/login":{"post":{"operationId":"UserController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/users/{id}":{"put":{"operationId":"UserController_update","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/change/password":{"put":{"operationId":"UserController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/change/email":{"put":{"operationId":"UserController_changeEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeEmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/profile":{"get":{"operationId":"UserController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/tikects":{"get":{"operationId":"TikectController_getTickets","parameters":[],"responses":{"200":{"description":""}},"tags":["Tikects"]}},"/tikects/{id}":{"get":{"operationId":"TikectController_getOneTicket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Tikects"]}},"/tikects/agents/{id}":{"get":{"operationId":"TikectController_geticketByAgents","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Tikects"]}},"/tikects/status/{status}":{"get":{"operationId":"TikectController_geticketByStatus","parameters":[{"name":"status","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Tikects"]}},"/constantees":{"post":{"operationId":"ConstanteController_createConstante","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConstanteDto"}}}},"responses":{"201":{"description":""}},"tags":["constantees"]},"get":{"operationId":"ConstanteController_init","parameters":[],"responses":{"200":{"description":""}},"tags":["constantees"]}},"/constantees/{id}":{"put":{"operationId":"ConstanteController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConstanteDto"}}}},"responses":{"200":{"description":""}},"tags":["constantees"]}},"/constantees/search":{"post":{"operationId":"ConstanteController_search","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConstanteDto"}}}},"responses":{"201":{"description":""}},"tags":["constantees"]}},"/constantees/search-first":{"post":{"operationId":"ConstanteController_findFirst","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConstanteDto"}}}},"responses":{"201":{"description":""}},"tags":["constantees"]}},"/constantees/order":{"get":{"operationId":"ConstanteController_getOrder","parameters":[],"responses":{"200":{"description":""}},"tags":["constantees"]}},"/constantees/order/reset":{"get":{"operationId":"ConstanteController_resetOrder","parameters":[],"responses":{"200":{"description":""}},"tags":["constantees"]}}},"info":{"title":"SIRMO API","description":"Système d'Identification RFID Mobile des conducteurs vehicules(Conducteurs) au Bénin","version":"1.0","contact":{}},"tags":[{"name":"cast","description":""}],"servers":[],"components":{"securitySchemes":{"token":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"UserDto":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"gender":{"type":"string","default":"MASCULIN"},"password":{"type":"string"},"email":{"type":"string"},"birth_date":{"format":"date-time","type":"string"},"phone":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}},"required":["firstname","lastname","gender","password","email","birth_date","phone","roles"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"ChangePasswordDto":{"type":"object","properties":{}},"ChangeEmailDto":{"type":"object","properties":{"old":{"type":"string"},"nevel":{"type":"string"}},"required":["old","nevel"]},"ConstanteDto":{"type":"object","properties":{"id":{"type":"number"},"nom":{"type":"string"},"valeur":{"type":"string"},"visibilite":{"type":"string"},"status":{"type":"boolean"},"description":{"type":"string"}},"required":["id","nom","valeur","visibilite","status","description"]}}}}